move motif stuff into separate file
